requirements.txt
flask==3.0.3
google-genai==1.2.0
python-dotenv==1.0.1
